Prisma Yoga Flow in Authorization: Implementing DeleteItem Permissions

+++ src/resolvers/Mutation.js
@@ -51,9 +51,16 @@ const Mutations = {
   async deleteItem(parent, args, context, info) {
     const where = { id: args.id };
     // 1. Find the Item
-    const item = await context.db.query.item({ where }, `{ id title }`);
+    const item = await context.db.query.item({ where }, `{ id title user { id }}`);
     // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === context.request.userId;
+    const hasPermissions = context.request.user.permissions.some(
+      permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
     
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have Permission to Delete this Item");
+    };
     // 3. Delete it
     return context.db.mutation.deleteItem({ where }, info);
   },
